pyproject.toml
[tool.poetry]
name = "fixbackend"
version = "0.1.0"
description = ""
authors = ["Some Engineering Inc."]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fastapi = "^0.101.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
fastapi-users = {extras = ["oauth", "sqlalchemy"], version = "^12.1.1"}
aiomysql = "^0.2.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.20"}
alembic = "^1.11.3"
pydantic-settings = "^2.0.3"
attrs = ">=23.1.0"
jinja2 = "^3.1.2"
boto3 = "^1.28.45"
async-lru = "^2.0.4"
arq = "^0.25.0"
fixcloudutils = { extras = ["redis", "arango"], version = "^1.13.1" }
prometheus-fastapi-instrumentator = ">=6.1.0"
websockets = "^11.0.3"
cryptography = "^41.0.6"
aiofiles = "^23.2.1"
pydantic-extra-types = "^2.2.0"
pycountry = "^23.12.11"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.1"
mypy = "^1.5.1"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
sqlalchemy-utils = "^0.41.1"
pymysql = "^1.1.0"
black = ">=23.7.0"
flake8 = ">=6.1.0"
coverage = {extras = ["toml"], version = "^7.3.1"}
pytest-cov = "^4.1.0"
httpx-ws = { git = "https://github.com/frankie567/httpx-ws.git", rev = "f9f2555603f40053f84adf6890e6cda245294b18" }
urllib3 = "<2" # poetry bug

# all type sheds
types-PyMySQL = "*"
types-pyOpenSSL = "*"
types-python-dateutil = "*"
types-pytz = "*"
types-PyYAML = "*"
types-requests = "*"
types-setuptools = "*"
types-ujson = "*"
types-urllib3 = "*"
types-aiofiles = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
fixbackend = "fixbackend.__main__:main"


[tool.pytest.ini_options]



[tool.coverage.paths]
source = ["fixbackend", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["fixbackend"]

[tool.coverage.report]
show_missing = true
fail_under = 80
